+++ crates/oxc_isolated_declarations/src/class.rs
@@ -253,14 +253,16 @@ impl<'a> IsolatedDeclarations<'a> {
         let mut elements = self.ast.vec();
         for (index, param) in function.params.items.iter().enumerate() {
             if param.accessibility.is_some() || param.readonly {
-                let type_annotation =
-                    if param.accessibility.is_some_and(TSAccessibility::is_private) {
-                        None
-                    } else {
-                        // transformed params will definitely have type annotation
-                        // SAFETY: `ast.copy` is unsound! We need to fix.
-                        unsafe { self.ast.copy(&params.items[index].pattern.type_annotation) }
-                    };
+                let type_annotation = if param
+                    .accessibility
+                    .is_some_and(TSAccessibility::is_private)
+                {
+                    None
+                } else {
+                    // transformed params will definitely have type annotation
+                    // SAFETY: `ast.copy` is unsound! We need to fix.
+                    unsafe { self.ast.copy(&params.items[index as u32].pattern.type_annotation) }
+                };
                 if let Some(new_element) =
                     self.transform_formal_parameter_to_class_property(param, type_annotation)
                 {
