+++ crates/oxc_linter/src/ast_util.rs
@@ -199,7 +199,7 @@ pub fn get_enclosing_function<'a, 'b>(
 
 /// Returns if `arg` is the `n`th (0-indexed) argument of `call`.
 pub fn is_nth_argument<'a>(call: &CallExpression<'a>, arg: &Argument<'a>, n: usize) -> bool {
-    let nth = &call.arguments[n];
+    let nth = &call.arguments[n as u32];
     nth.span() == arg.span()
 }
